Refactor: fetchUtil를 debateRoom 내부 Util로 이동
